SCRD-2407 Magnum service fixes
Set the project name to "services" because we don't actually
create a magnum project.

Restore removed configuration items as they are included in the
playbooks. While the values are duplicate it is conceivable that
they could be set differently in future.

Add a transport_url to replace deprecated rabbitmq configs.

Also, the [keystone_authtoken] section is deprecated in favor of
[keystone_auth] for magnum-conductor so we need to update magnum.conf.j2
accordingly.

See https://review.openstack.org/#/c/274227/

Change-Id: Ic9edc1109c627d2d9932104022f1ebec1d07aea4