Use new local internal cert fact (bsc#1088501)
Previously, running magnum deploy and upgrade playbooks
could not be done when limited to a subset of hosts not
including TLS_CA--first-member. This was due to how the
name of the internal CA cert was derived in the playbook.
With modifications to tls-ansible, the cert name is now
gathered when tls-trust is run on TLS_CA--first-member, and
then stored in /etc/ansible/facts.d on all hosts so that it
can be used later. I have removed the previous means for
getting the filename here since it is no longer required,
and simplified how the fact is set here.

In the course of investigating the best solution for this
bug I encountered two bits of configuration no longer used:

- The magnum_conf variable no longer exists and the config
parameter to _write_conf.yml was never used as far as I can
tell. Remove it.

- The add_ca_cert_file option in magnum.conf is only
applicable to the earlier in-house patched version of
Magnum - remove it and the cluster section from the template
as it does nothing. (The cluster section does exist upstream,
but we do not include any configuration items from it here.)

Depends-On: I8c596a47aa1083940872b0e059dd9ad32d1a8b7b

Change-Id: I07fcd039761177a94e3d203e9fff020c046915ca